Use vectors in Camera interface
Using vectors wraps up the relation between the coordinates and makes it also
easier to pass actual vectors as arguments. This is especially tedious when the
vector to pass as argument is a complex expression. The obvious drawback of
this change is, that to use this with individual components one now has to
assemble them into a vector by adding [] first.